a11y: Fix crash in treeview
This is a band-aid fix. The master branch has a complete rework of the
treeview a11y code. Let's hope this will cause most crashes to magically
disappear. This code has been indexing by random memory on the stack for
a long while and things didn't crash, so let's hope for the best.

https://bugzilla.gnome.org/show_bug.cgi?id=663322
https://bugzilla.gnome.org/show_bug.cgi?id=664137